+++ src/Algos/Projection.cpp
@@ -147,12 +147,7 @@ bool NOMAD::Projection::runImp()
 {
     bool projectionOk = true;
 
-    // Return value: found better - unused.
-    // TODO Ensure OPPORTUNISM is off here
     evalTrialPoints(this);
-
-    // TODO - postprocessing?
-
 
     return projectionOk;
 }
@@ -172,9 +167,6 @@ void NOMAD::Projection::projectPoint(const NOMAD::EvalPoint& oraclePoint)
 
     // STD projected point
     stdProjectedPoint(oraclePoint);
-    // TODO compute this otherwise, if needed
-    //auto f = bestEvalPoint.getF(evalType);
-    //auto h = bestEvalPoint.getH(evalType);
 
     // Try pertubation
     std::string subStepName = "Projection candidate";
@@ -236,15 +228,6 @@ void NOMAD::Projection::projectPoint(const NOMAD::EvalPoint& oraclePoint)
     trySet.clear();
     keep.clear();
 
-
-    // Evaluate projection trial points
-    // in the dynamic (quad or sgtelib) model
-    // TODO Analyse from NOMAD 3 and see if we can do something similiar
-    // in NOMAD 4. It may not be worth it, it seems more like an
-    // issue of sorting the points accorting to a SgtelibModel, and
-    // that would be better done in the EvaluatorControl.
-    //evaluateProjectionTrialPoints(trySet, ev, keep, bestEvalPoint);
-
 }
 
 
@@ -272,13 +255,11 @@ void NOMAD::Projection::stdProjectedPoint(const NOMAD::EvalPoint& oraclePoint)
 
     if (nullptr != _mesh)
     {
-        // TODO: Use mesh and frame center from IterationUtils
         // Project the point on the mesh
         xTry = _mesh->projectOnMesh(xTry, *_frameCenter);
     }
     NOMAD::EvalPoint evalPoint(xTry);
 
-    // TODO This code is based on NOMAD 3's Sgtelib_Model code.
     // The goal here is to evaluate points according to the SgtelibModel.
     // This may not have its place in the Projection class.
     bool doInsert = true;
@@ -426,7 +407,7 @@ void NOMAD::Projection::doGreedySelection(const NOMAD::EvalPoint &oraclePoint,
             double Q = Ds[i]-lambda*Dr[i];
             if ( Q > Q_max )
             {
-                inew = i;
+                inew = (int)i;
                 Q_max = Q;
             }
         }
